% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_change.R
\name{abundance_change}
\alias{abundance_change}
\title{Species Abundance Changes}
\usage{
abundance_change(
  df,
  time.var,
  species.var,
  abundance.var,
  replicate.var = NULL,
  reference.time = NULL
)
}
\arguments{
\item{df}{A data frame containing time, species, and abundance columns and an
optional columns of replicates.}

\item{time.var}{The name of the time column.}

\item{species.var}{The name of the species column.}

\item{abundance.var}{The name of the abundance column.}

\item{replicate.var}{The name of the optional replicate column. If specified,
replicate must be unique within the dataset and cannot be nested within
treatments or blocks.}

\item{reference.time}{The name of the optional time point that all other time
points should be compared to (e.g. the first year of data). If not
specified, each comparison is between consecutive time points (the first
and second year, second and third year, etc.)}
}
\value{
The `abundance_change()` function returns a data frame with a subset of
  the following columns:
- **replicate.var**: A column with the specified `replicate.var`, if it is
  specified.
- **time.var**: A column with the specified `time.var` and a second column,
  with '2' appended to the name. Time is subtracted from time2.
- **species.var**: A column with the specified `species.var`.
- **change**: A numeric column of the change in abundance between time
  points. A positive value occurs when a species increases in abundance over
  time, and a negative value when a species decreases in abundance over time.#' @references Avolio et al. Submitted
}
\description{
Calculates the abundance change for species in a replicate
  between two time points. Changes are on abundance values provided, if
  relative data is used, then changes in relative abundance will be
  calculated.
}
\examples{
data(pplots)
# Without replicates
df <- subset(pplots, plot == 25)
abundance_change(df = df,
                 species.var = "species",
                 abundance.var = "relative_cover",
                 time.var = "year")

# With replicates
df <- subset(pplots, year < 2004 & plot \%in\% c(6, 25, 32))
abundance_change(df = df,
                 species.var = "species",
                 abundance.var = "relative_cover",
                 replicate.var = "plot",
                 time.var = "year")
                 
# With reference year
df <- subset(pplots, year < 2005 & plot \%in\% c(6, 25, 32))
abundance_change(df = df,
                 species.var = "species",
                 abundance.var = "relative_cover",
                 replicate.var = "plot",
                 time.var = "year",
                 reference.time = 2002)
}
