% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_rank_abundance}
\alias{add_rank_abundance}
\title{Add relative abundance ranks and cumulative abundance}
\usage{
add_rank_abundance(df, species.var, abundance.var)
}
\arguments{
\item{df}{A data frame containing a single record per species with its abundance}

\item{abundance.var}{The name of the abundance column}
}
\value{
The add_rank_abundance function returns a data frame with the following
  additional column:
  \itemize{
    \item{rank: }{}
    \item{relrank: }{A numeric column with the species rank divided
  by the maximum rank in the group; a rank of 1 indicates the species was the
  least abundant.}
    \item{cumabund: }{}
  }
}
\description{
Rank species by abundance within the specified grouping. The
  rank of the lead abundant species is 1 and the most abundant species has
  rank 1/S, where S is the number of species in the group.
}
