% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{calComDat}
\alias{calComDat}
\title{Convert from a longform abundance dataframe to a year x species dataframe.}
\usage{
calComDat(data1, species, year, abundance)
}
\arguments{
\item{data1}{A dataframe containing year, species and abundance columns}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
comdat A dataframe of species abundances x year
}
\description{
Convert from a longform abundance dataframe to a year x species dataframe.
}

