% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{calnullVR}
\alias{calnullVR}
\title{A function to calculate a null variance ratio from longform data using a temporal modification of the Torus translation}
\usage{
calnullVR(data1, species, year, abundance)
}
\arguments{
\item{data1}{A dataframe containing year, rep, species and abundance columns}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
randVR A variance ratio calculated from a randomized community matrix in which species autocorrelation has been maintained via a Torus translation
}
\description{
A function to calculate a null variance ratio from longform data using a temporal modification of the Torus translation
}

