% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{calrealVR}
\alias{calrealVR}
\title{A function to calculate the variance ratio within multiple replicates}
\usage{
calrealVR(data1, year, sitesubplot, species, abundance)
}
\arguments{
\item{data1}{A dataframe containing year, sitesubplot, species and abundance columns}

\item{year}{The name of the year column from data1}

\item{sitesubplot}{The name of the replicate column from data1}

\item{species}{The name of the species column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
output A dataframe containing sitesubplot and VR (the variance ratio)
}
\description{
calrealVR a dataframe of replicates with their variance ratio values
}

