% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variance_ratio.R
\name{calVR}
\alias{calVR}
\title{A function to calculate the variance ratio}
\usage{
calVR(comdat)
}
\arguments{
\item{comdat}{A community dataframe}
}
\value{
var.ratio The variance ratio of the community
}
\description{
A function to calculate the variance ratio
}

