% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{calVR_longformdata}
\alias{calVR_longformdata}
\title{A function to calculate the variance ratio from a longform dataframe}
\usage{
calVR_longformdata(data1, species, year, abundance)
}
\arguments{
\item{data1}{A dataframe containing year, rep, species and abundance columns}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
var.ratio The variance ratio of the community
}
\description{
A function to calculate the variance ratio from a longform dataframe
}

