% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{calVRs}
\alias{calVRs}
\title{A function to calculate the variance ratio within multiple replicates}
\usage{
calVRs(data1, rep, species, year, abundance)
}
\arguments{
\item{data1}{A dataframe containing year, rep, species and abundance columns}

\item{rep}{The name of the replicate column from data1}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
output A dataframe containing rep and VR (the variance ratio)
}
\description{
calVRs a dataframe of replicates with their variance ratio values
}

