% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_args}
\alias{check_args}
\title{Check for errors in the application of arguments and input data
for all *_difference functions.}
\usage{
check_args(df, time.var = NULL, species.var, abundance.var,
  replicate.var = NULL, treatment.var = NULL, pool = FALSE,
  block.var = NULL, reference.time = NULL, reference.treatment = NULL)
}
\description{
Check for errors in the application of arguments and input data
for all *_difference functions.
}
