% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_multispp}
\alias{check_multispp}
\title{Utility function to stop calculations if only one species occurs in at least one replicate}
\usage{
check_multispp(df, species.var, replicate.var)
}
\arguments{
\item{df}{A dataframe containing time.var, species.var and abundance.var columns}

\item{species.var}{The name of the species column from df}

\item{replicate.var}{The name of the replicate column from df}
}
\description{
Utility function to stop calculations if only one species occurs in at least one replicate
}
