% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_numeric}
\alias{check_numeric}
\title{Utility to check for numeric abundance and time variables}
\usage{
check_numeric(df, time.var, abundance.var)
}
\arguments{
\item{df}{A dataframe containing time.var, species.var, and replicate.var columns}

\item{time.var}{The name of the time column from df}

\item{abundance.var}{The name of the replicate column from df}
}
\description{
Utility to check for numeric abundance and time variables
}
