% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_sppvar}
\alias{check_sppvar}
\title{Utility function to stop calculations if the species never change in a replicate}
\usage{
check_sppvar(comdat)
}
\arguments{
\item{comdat}{A community dataframe}
}
\description{
Utility function to stop calculations if the species never change in a replicate
}
