% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/codyn-package.r
\docType{package}
\name{codyn}
\alias{codyn}
\alias{codyn-package}
\title{Ecological Community Dynamics}
\description{
Temporal diversity indices and community stability metrics for ecologists.
}
\details{
The functions in codyn implement metrics that are explicitly temporal, and include the option to  calculate them over multiple replicates. Functions fall into two categories: temporal diversity indices and community stability metrics. The diversity indices in codyn are temporal analogs to traditional diversity indices such as richness and rank-abundance curves. Specifically, codyn includes functions to calculate species turnover, mean rank shifts and lags in community similarity between time points. The community stability metrics in codyn calculate overall stability and patterns of species covariance and synchrony over time. Finally, codyn contains vignettes that describe methods and reproduce figures from published papers to help users contextualize and apply functions to their own data.
}
\author{
Lauren M. Hallett
}

