% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rate_change.R
\name{community_distance}
\alias{community_distance}
\title{A function to calculate euclidean community distance over time within multiple replicates.}
\usage{
community_distance(df, time.var = "time", species.var = "species",
  abundance.var = "abundance", replicate.var = NA)
}
\arguments{
\item{df}{A dataframe containing replicate, time, species and abundance columns.}

\item{time.var}{The name of the time column from df}

\item{species.var}{The name of the species column from df}

\item{abundance.var}{The name of the abundance column from df}

\item{replicate.var}{The name of the replicate column from df. Defaults to NA.}
}
\value{
output a data frame with the euclidean distance between communities by replicate and interval
}
\description{
This is an analysis of differences in species composition between samples at increasing time
lags. It measures the rate of directional change in community composition. First, a
triangular dissimilarity matrix is calculated using Euclidean distance. Then, the Euclidean
distance values are plotted against time lags, and output as a data frame.
}

