% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{curve_dissim}
\alias{curve_dissim}
\title{Rank-Abundance Curve Dissimilarity
A function to calculate the curve difference between two communities, given the
empirical relative rank by abundance curve as a \code{stepfun}.}
\usage{
curve_dissim(sf, sf2)
}
\arguments{
\item{sf}{The curve for the first community.}

\item{sf2}{The curve for the second community.}
}
