% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_torus_translation.R
\name{cyclic_shift_onerep}
\alias{cyclic_shift_onerep}
\title{A function to calculate a non-S3 cyclic shift on one replicate}
\usage{
cyclic_shift_onerep(df, time.var, species.var, abundance.var, FUN, bootnumber)
}
\arguments{
\item{df}{A data frame containing time, species and abundance columns and an optional column of replicates}

\item{time.var}{The name of the time column}

\item{species.var}{The name of the species column}

\item{abundance.var}{The name of the abundance column}

\item{FUN}{A function to calculate on the null community}

\item{bootnumber}{The number of null model iterations returned}
}
\value{
out A vector of  test statistics calculated on the null community
}
\description{
A function to calculate a non-S3 cyclic shift on one replicate
}
