% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_shift.R
\name{df_intersect}
\alias{df_intersect}
\title{Create intersected data frames}
\usage{
df_intersect(df1, df2, dataname = "species")
}
\arguments{
\item{df1}{A dataframe}

\item{df2}{A dataframe}

\item{dataname}{The name of the column on which the two datasets will be joined and intersected}
}
\description{
Create intersections.
}
