% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity_change.R
\name{dissimilarity_change}
\alias{dissimilarity_change}
\title{Bray-Curtis dissimilarity of replicates between and within time periods}
\usage{
dissimilarity_change(df, time.var, species.var, abundance.var, replicate.var,
  treatment.var = NULL)
}
\arguments{
\item{df}{A data frame containing time, species, abundance and replicate columns and an optional column of treatment}

\item{time.var}{The name of the time column}

\item{species.var}{The name of the species column}

\item{abundance.var}{The name of the abundance column}

\item{replicate.var}{The name of the replicate column}

\item{treatment.var}{the neame of the optional treatment column}
}
\value{
The multivariate_change function returns a data frame with the following attributes:
\itemize{
 \item{time.var: }{A characteric column that has the first of two time periods that are being compared.}
 \item{time.var2: }{A characteric column that has the second of two time periods that are being compared.}
 \item{BC_between_change: }{A numeric column that is the average pairwise Bray-Curtis dissimilarity of replicates in two consecutive time periods. 0 - The communities are similar over time. 1 - The communities are changing over time.}
 \item{BC_within_change: }{A numeric column that is change between two time periods in the average pairwise Bray-Curtis dissimilarity of replicates within a time period. A negative number indicates that time.var2 has varaiblity in community composition than time.var}
 \item{treatment.var: }{A column that has same name and type as the treatment.var column, if treatment.var is specified.}
}
}
\description{
Calculates the average changes in Bray-Curtis dissimilarity of replicates between two consecutive time periods and within time periods. Between changes the average Bray-Curtis dissimilaritiy of all pairwise comparisions of replicates between two consecutive time periods. This is a measure of how dissimilar the community composition of two time periods is. Bray-Curtis dissimilarity ranges from 0-1, where 0 are identical communities, and 1 and completelty different communiites. Within change is derived in two steps. First, the average Bray-Curtis dissimilarity of all replicates within a time period is calculated. This is a measure of how homogenous a community is within a time step is. Then, these averages are compared between two consecutive time periods.
}
\examples{
data(pplots)
#With treatment
dissimilarity_change(pplots,
                    time.var="year", 
                    replicate.var = "plot", 
                    treatment.var = "treatment", 
                    species.var = "species", 
                    abundance.var = "relative_cover")

#Without treatment
df <- subset(pplots, treatment == "N1P0")
dissimilarity_change(df, 
                    time.var="year", 
                    replicate.var = "plot", 
                    species.var = "species", 
                    abundance.var = "relative_cover")
}
\references{
Avolio et al. 2015; Avolio et al. OUR PAPER, Mari Anderson?
}
