% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity_difference.R
\name{dissimilarity_difference}
\alias{dissimilarity_difference}
\title{Bray curtis dissimilarity differences in between and within treatments}
\usage{
dissimilarity_difference(df, time.var = NULL, species.var, abundance.var,
  replicate.var, treatment.var)
}
\arguments{
\item{df}{A data frame containing an optional time column, species, abundance and replicate, and treatment columns}

\item{time.var}{The name of the optional time column}

\item{species.var}{The name of the species column}

\item{abundance.var}{The name of the abundance column}

\item{replicate.var}{The name of the replicate column}

\item{treatment.var}{the name of the treatment column}
}
\value{
The multivariate_difference function returns a data frame with the following attributes:
\itemize{
 \item{treatment.var: }{A column that has same name and type as the treatment.var column, if treatment.var is specified.}
 \item{treatment.var2: }{A column that has the same type as the treatment.var column, and is named treatment.var with a 2 appended to it.}
 \item{BC_between_diff: }{A numeric column that is the average pairwise Bray-Curtis dissimilarity of all pairwise comparisons of replicates in both treatments.}
 \item{BC_within_diff: }{A numeric column that is difference in the average Bray-Curtis dissimilarity of all pairwise comparisions of replicates within each treatment.}
 \item{time.var: }{A characteric column that has the same name and type as the time.var column, if specified.}
}
}
\description{
Calculates the average difference in Bray-Curtis dissimilarity between treatments and within treatments at a single point in time. Between differences is the average Bray-Curtis dissimilaritiy of all pairwise comparisions of replicates between two treatments. Within difference is the differecnes between two treatments of average dissimilarity of all replicates within each treatment.
}
\examples{
 
data(pplots)
#Without time
df <- subset(pplots, year == 2002)
dissimilarity_difference(df, 
                        replicate.var = "plot", 
                        treatment.var = "treatment", 
                        species.var = "species", 
                        abundance.var = "relative_cover")
#With time
dissimilarity_difference(pplots, 
                        time.var = "year", 
                        replicate.var = "plot", 
                        species.var = "species", 
                        abundance.var = "relative_cover", 
                        treatment.var = "treatment")
}
\references{
Our Avolio et al. paper, Avolio et al. 2015, Marti Anderson?
}
