% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_structure.R
\name{EQ}
\alias{EQ}
\title{Utility function to calculate EQ evenness from Smith and Wilson 1996}
\usage{
EQ(x)
}
\arguments{
\item{x}{Vector of abundance of each species
If all abundances are equal it returns a 1}
}
\description{
Utility function to calculate EQ evenness from Smith and Wilson 1996
}
