% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Evar}
\alias{Evar}
\title{Evar}
\usage{
Evar(x, S = length(x))
}
\arguments{
\item{x}{the vector of abundances of each species}

\item{S}{the number of species in the sample}
}
\description{
A utility function to calculate Evar from Smith and Wilson 1996
}
