% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genRand}
\alias{genRand}
\title{A function to generate a community dataframe with a random start time for each species}
\usage{
genRand(comdat)
}
\arguments{
\item{comdat}{A community dataframe}
}
\value{
rand.use A randomized community dataframe
}
\description{
A function to generate a community dataframe with a random start time for each species
}

