% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rate_change.R
\name{get_intervals}
\alias{get_intervals}
\title{Get lagged intervals
Returns lagged intervals of community change within one replicate.}
\usage{
get_intervals(df, time.var = "time", species.var = "species",
  abundance.var = "abundance")
}
\arguments{
\item{df}{data frame to compute the slope of community change for}

\item{time.var}{The name of the time column from df}

\item{species.var}{The name of the species column from df}

\item{abundance.var}{The name of the abundance column from df}
}
\value{
a data frame containing of time lags by species distances
}
\description{
Get lagged intervals
Returns lagged intervals of community change within one replicate.
}

