% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{get_slope}
\alias{get_slope}
\title{A function that returns the slope of community change within one replicate.}
\usage{
get_slope(data1, year = "year", species = "species",
  abundance = "abundance")
}
\arguments{
\item{year}{The name of the year column from data1}

\item{species}{The name of the species column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
a slope of year lags by species distances
}
\description{
A function that returns the slope of community change within one replicate.
}

