% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankshifts.R
\name{getintersect}
\alias{getintersect}
\title{Create intersected data frames}
\usage{
getintersect(d1, d2, dataname = "species")
}
\arguments{
\item{d1}{A dataframe}

\item{d2}{A dataframe}

\item{dataname}{The name of the column on which the two datasets will be joined and intersected}
}
\description{
Create intersections.
}

