% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getturnover}
\alias{getturnover}
\title{A function to calculate species turnover between two years}
\usage{
getturnover(d1, d2, species = "species", metric = "turnover")
}
\arguments{
\item{d1}{A dataframe containing a species column from one year}

\item{d2}{A dataframe containing a species column from the following year}

\item{species}{The name of the species column in d1 and d2}

\item{metric}{The turnover metric to return; the default, turnover, returns summed appearances and disappearances relative to total species richness across both years
appearance returns the number of appearances in the second year relative to total species richness across both years
disappearance returns the number of disappearances in the second year relative to the total species richness across both years}
}
\value{
output The specificed turnover metric
}
\description{
A function to calculate species turnover between two years
}

