% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_change.R
\name{lag_i}
\alias{lag_i}
\title{Get lagged values from a distance matrix
Get lagged values from distance matrix at value i}
\usage{
lag_i(i, DM, rownums, colnums)
}
\arguments{
\item{i}{the index of the matrix to lag}

\item{DM}{the distance matrix from which lagged values are drawn}

\item{rownums}{number of rows in the distance matrix}

\item{colnums}{number of columns in the distance matrix}
}
\value{
the lagged values
}
\description{
Get lagged values from a distance matrix
Get lagged values from distance matrix at value i
}
