% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_change.R
\name{lagged_distances}
\alias{lagged_distances}
\title{Get lagged distances for a single replicate}
\usage{
lagged_distances(df, time.var, species.var, abundance.var)
}
\arguments{
\item{df}{data frame to compute the slope of community change for}

\item{time.var}{The name of the time column from df}

\item{species.var}{The name of the species column from df}

\item{abundance.var}{The name of the abundance column from df}
}
\value{
a data frame containing of time lags by species distances
}
\description{
Returns a data frame with two columns, interval and distance. The interval is
the number of time steps between two communities, while distance is the
euclidean distance of community change within one replicate lagged across invtervals.
}
