% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_change.R
\name{lagged_slope}
\alias{lagged_slope}
\title{Get slope
Returns the slope of community change within one replicate.}
\usage{
lagged_slope(df, time.var, species.var, abundance.var)
}
\arguments{
\item{df}{data frame to compute the slope of community change for}

\item{time.var}{The name of the time column from df}

\item{species.var}{The name of the species column from df}

\item{abundance.var}{The name of the abundance column from df}
}
\value{
a slope of time lags by species distances
}
\description{
Get slope
Returns the slope of community change within one replicate.
}
