% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rankshifts.R
\name{meanrankshift}
\alias{meanrankshift}
\title{Mean Rank Shifts}
\usage{
meanrankshift(df, time.var = "year", species.var = "species",
  abundance.var = "abundance", replicate.var = as.character(NA))
}
\arguments{
\item{df}{dataframe of Community dataset. Must be in 'long' format.}

\item{time.var}{The time variable}

\item{species.var}{The species variable}

\item{abundance.var}{The abundance variable}

\item{replicate.var}{The replication variable}
}
\value{
a dataframe, showing years compared
}
\description{
Quantify shifts in species rank order
}
\details{
Mean rank shift (MRS) quantifies relative changes in species rank abundances by taking the sum difference of species ranks of consecutive pairs of years.
MRS is an indicator of shifts in relative abundance over time.
}
\references{
Collins, Scott L., Katharine N. Suding, Elsa E. Cleland, Michael Batty, Steven C. Pennings, Katherine L. Gross, James B. Grace, Laura Gough, Joe E. Fargione, and Christopher M. Clark. “Rank Clocks and Plant Community Dynamics.” Ecology 89, no. 12 (2008): 3534–41.
}

