% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{nullVRCI}
\alias{nullVRCI}
\title{A function to generate lower 2.5\% CI, upper 97.5\% CI and mean null VR values}
\usage{
nullVRCI(data1, species, year, abundance, bootnumber)
}
\arguments{
\item{data1}{A dataframe containing year, rep, species and abundance columns}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
output A dataframe containing nullVRCIlow, nullVRCIhigh and nullVRmean
         low is the lower 0.025
}
\description{
A function to generate lower 2.5\% CI, upper 97.5\% CI and mean null VR values
}

