% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pairwise_turnover}
\alias{pairwise_turnover}
\title{A function to calculate species turnover between two years within a rep}
\usage{
pairwise_turnover(data1, rep, species, year, abundance)
}
\arguments{
\item{data1}{A dataframe containing year, rep, species and abundance columns}

\item{rep}{The name of the replicate column from data1}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
turnover The species turnover (appearances + disappearances) relative to the total species richness observed across both years
}
\description{
A function to calculate species turnover between two years within a rep
}

