% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_shift.R
\name{rank_onerep}
\alias{rank_onerep}
\title{Function for calculating mean rank shifts}
\usage{
rank_onerep(df, time.var, species.var, abundance.var)
}
\arguments{
\item{df}{dataframe of Community dataset. Must be in 'long' format.}

\item{time.var}{The time variable}

\item{species.var}{The species variable}

\item{abundance.var}{The abundance variable}
}
\value{
a dataframe, showing years compared
}
\description{
This is a function that calculates mean rank shifts
}
