% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{stability_onerep}
\alias{stability_onerep}
\title{A function to calculate species synchrony over time within one replicate}
\usage{
stability_onerep(data1, x)
}
\arguments{
\item{data1}{A dataframe containing x column}

\item{x}{The column to calculate stability on}
}
\value{
Stability of x, calculated as the mean/sd
}
\description{
A function to calculate species synchrony over time within one replicate
}

