% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{synch_onerep}
\alias{synch_onerep}
\title{A function to calculate species synchrony over time within one replicate}
\usage{
synch_onerep(data1, species, year, abundance)
}
\arguments{
\item{data1}{A dataframe containing rep, year, species and abundance columns}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
output The degree of species synchrony, where 1 is perfect synchrony and 0 is perfect asynchrony
}
\description{
A function to calculate species synchrony over time within one replicate
}

