% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{timelag}
\alias{timelag}
\title{Function for time-lag analysis or directional change in temporal Community composition data.}
\usage{
timelag(comm_data)
}
\arguments{
\item{comm_data}{Community dataset. Must be in 'wide' format}

\item{year}{Year variable. Must be the first column of comm_data}
}
\value{
linear model coefficients
}
\description{
This is a function that calculates the slope of directional change.
}

