% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_ratio.R
\name{variance_ratio_longformdata}
\alias{variance_ratio_longformdata}
\title{A function to calculate the variance ratio from a longform dataframe}
\usage{
variance_ratio_longformdata(df, time.var, species.var, abundance.var)
}
\arguments{
\item{df}{A dataframe containing time.var, replicate.var, species.var and abundance.var columns}

\item{time.var}{The name of the time.var column from df}

\item{species.var}{The name of the species.var column from df}

\item{abundance.var}{The name of the abundance.var column from df}
}
\value{
var.ratio The variance ratio of the community
}
\description{
A function to calculate the variance ratio from a longform dataframe
}
