% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_ratio.R
\name{variance_ratio_matrixdata}
\alias{variance_ratio_matrixdata}
\title{A function to calculate the variance ratio}
\usage{
variance_ratio_matrixdata(comdat)
}
\arguments{
\item{comdat}{A community dataframe}
}
\value{
var.ratio The variance ratio of the community
}
\description{
A function to calculate the variance ratio
}
