% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variance_ratio.R
\name{varianceratio}
\alias{varianceratio}
\title{Variance Ratio}
\usage{
varianceratio(df, time.var = "year", species.var = "species",
  abundance.var = "abundance", bootnumber, replicate.var = NA, li = 0.025,
  ui = 0.975, average.replicates = TRUE)
}
\arguments{
\item{df}{A dataframe containing time.var, species.var and abundance.var columns and an optional replicate.var column}

\item{time.var}{The name of the time column from df}

\item{species.var}{The name of the species column from df}

\item{abundance.var}{The name of the abundance column from df}

\item{bootnumber}{The number of null model iterations used to calculated CIs}

\item{replicate.var}{The name of the replicate column from df}

\item{li}{The lower confidence interval, defaults to lowest 2.5\% CI}

\item{ui}{The upper confidence interval, defaults to upper 97.5\% CI}

\item{average.replicates}{If true returns VR and CI averaged across reps; if false returns VR and CI for each rep}
}
\value{
output A dataframe containing the replicate.var name, VR  nullVRCIlow, nullVRCIhigh and nullVRmean
         VR is the actual variance ratio
         nullVRCIow is the 0.025 CI
         nullVRCIhigh is the 0.975 CI
         nullVRmean is the mean variance ratio calculated on null communities
}
\description{
Computes the ratio of the variance of aggregate species abundances in a community to the sum of the variances of individual, component species. A variance ratio = 1 indicates that species do not covary,  a variance ratio > 1 indicates predominately positive covariance among species and a variance ratio < 1 indicates predominately negative covariance (Schulter 1984).

Includes a null modeling option to test if the variance ratio significantly differs from 1. The null community is created by randomly selecting different starting points for each species' time series, which generates a community in which species abundances vary independently but within-species autocorrelation is maintained (Hallett et al. 2014). This randomization is repeated a user-specific number of times and confidence intervals are reported for the resultant null distribution of variance ratios. If the dataframe includes multiple replicates, the variance ratios for the actual and null communities are averaged within each iteration unless specified otherwise.
}
\details{
\deqn{ VR = \frac{Var(C)}{\sum_{i}^{N} Var(x_i)}}
}

