% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variance_ratio.R
\name{varianceratio_matrixdata}
\alias{varianceratio_matrixdata}
\title{A function to calculate the variance ratio}
\usage{
varianceratio_matrixdata(comdat)
}
\arguments{
\item{comdat}{A community dataframe}
}
\value{
var.ratio The variance ratio of the community
}
\description{
A function to calculate the variance ratio
}

