% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_change.R
\name{abundance_change}
\alias{abundance_change}
\title{Species Abundance Changes}
\usage{
abundance_change(df, time.var, species.var, abundance.var,
  replicate.var = NULL)
}
\arguments{
\item{df}{A data frame containing time, species, and abundance columns and an
optional column of replicates}

\item{time.var}{The name of the time column}

\item{species.var}{The name of the species column}

\item{abundance.var}{The name of the abundance column}

\item{replicate.var}{The name of the optional replicate column}
}
\value{
The abundance_change function returns a data frame with the following
  fields:
\itemize{
 \item{replicate.var: }{A column with the specified replicate.var, if it is specified.}
 \item{time.var: }{A column with the specified time.var and a second column, with '2'
 appended to the name, giving the time of the subtracted abundance.}
 \item{species.var: }{A column with the specified species.var.}
 \item{change: }{A numeric column of the change in abundance between consecutive
 timepoints. A postive value occurs when a species increases in abundnace
 over time, and a negative value when a species decreases in abundance over
 time.}
}
}
\description{
Calculates the abundance change for species in a replicate
  between two consecutive time points.
}
\examples{
data(pplots)
# Without replicates
df <- subset(pplots, plot == 25)
abundance_change(df = df,
                 species.var = "species",
                 abundance.var = "relative_cover",
                 time.var = "year")

# With replicates
df <- subset(pplots, year < 2004 & plot \%in\% c(6, 25, 32))
abundance_change(df = df,
                 species.var = "species",
                 abundance.var = "relative_cover",
                 replicate.var = "plot",
                 time.var = "year")
}
\references{
Avolio et al. OUR PAPER
}
