% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{calComDat}
\alias{calComDat}
\title{A function to covert from a longform abundance dataframe to a species x year dataframe}
\usage{
calComDat(data1, species, year, abundance)
}
\arguments{
\item{data1}{A dataframe containing year, species and abundance columns}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
comdat A dataframe of species abundances x year
}
\description{
A function to covert from a longform abundance dataframe to a species x year dataframe
}

