% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_args}
\alias{check_args}
\title{Checking for errors in arguments}
\usage{
check_args(df, time.var = NULL, species.var, abundance.var,
  replicate.var = NULL, treatment.var = NULL, pool = FALSE,
  block.var = NULL, reference.time = NULL, reference.treatment = NULL)
}
\arguments{
\item{df}{The name of the dataframe}

\item{time.var}{The name of the optional time column}

\item{species.var}{The name of the species column}

\item{abundance.var}{The name of the abundance column}

\item{replicate.var}{The name of the replicate column}

\item{treatment.var}{The name of the optional treatment column}

\item{pool}{The name the optional pooling approach}

\item{block.var}{The name of the optional block column}

\item{reference.time}{The name of the optional reference time}

\item{reference.treatment}{The name of the optional reference treatment}
}
\description{
Check for errors in the application of arguments and input data
  for all \code{*_change} and \code{*_difference} functions.
}
