% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_names}
\alias{check_names}
\title{check names of data frames}
\usage{
check_names(given, data)
}
\arguments{
\item{given}{Vector of variable names as supplied by user}

\item{data}{Data frame containing variables}
}
\description{
check names of data frames
}
