% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_single}
\alias{check_single}
\title{Utility function to ensure only a single record exists for a given species
within one replicate, for one time point.}
\usage{
check_single(df, time.var, species.var, replicate.var)
}
\arguments{
\item{df}{A dataframe containing time.var, species.var, and replicate.var columns}

\item{time.var}{The name of the time column from df}

\item{species.var}{The name of the species column from df}

\item{replicate.var}{The name of the replicate column from df}
}
\description{
Utility function to ensure only a single record exists for a given species
within one replicate, for one time point.
}
