% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/codyn-package.r
\docType{package}
\name{codyn}
\alias{codyn}
\alias{codyn-package}
\title{codyn title goes here}
\description{
codyn description goes here
}
\details{
include details here
}
\author{
Lauren Hallett
}

