% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/codyn-package.r
\docType{data}
\name{collins08}
\alias{collins08}
\title{Konza data from Collins et. all 2008}
\format{A data frame with 2058 rows and 4 variables}
\source{
Scott L. Collins 1987. Interaction of Disturbances in Tallgrass Prairie: A Field Experiment.
Ecology 68:1243–1250. http://dx.doi.org/10.2307/1939208
}
\usage{
data(collins08)
}
\description{
A dataset of tallgrass prairie plant composition at one annually burned and one unburned
site over time at the Konza Prairie LTER, Manhattan Kansas (Collins et al. 2008).
}
\details{
\itemize{
  \item replicate The replicate type, i.e. "burned", "unburned"
  \item year The sampling frequency
  \item species The sampled species
  \item abundance The measure of species abundance
}
}
\keyword{datasets}

