% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/community_stability.R
\name{community_stability}
\alias{community_stability}
\title{Community Stability}
\usage{
community_stability(df, time.var = "year", abundance.var = "abundance",
  replicate.var = NA)
}
\arguments{
\item{df}{the data frame to use in the calculation}

\item{time.var}{The name of the time column from df}

\item{abundance.var}{The name of the abundance column from df}

\item{replicate.var}{The name of the replicate column from df}
}
\value{
The stability of community abundance, calculated as mean/standard deviation returns a single number or a dataframe with the following attributes:
\itemize{
 \item{replicate}{replicate name}
 \item{stability}{community stability for replicate}
 }
}
\description{
A function to calculate community stability over time
}
\details{
The input dataframe needs to contain columns for time and abundance;
}

