% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/community_stability.R
\name{community_stability}
\alias{community_stability}
\title{A function to calculate community stability over time}
\usage{
community_stability(df, time.var = "year", abundance.var = "abundance",
  replicate.var = as.character(NA))
}
\arguments{
\item{df}{the data frame to use in the calculation}

\item{time.var}{The name of the time column from df}

\item{abundance.var}{The name of the abundance column from df}

\item{replicate.var}{The name of the replicate column from df}
}
\value{
output The stability of community abundance, calculated as mean/standard deviation
}
\description{
A function to calculate community stability over time
}

