% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{community_stability}
\alias{community_stability}
\title{A function to calculate community stability over time}
\usage{
community_stability(data1, replicate = "replicate", year = "year",
  abundance = "abundance")
}
\arguments{
\item{replicate}{The name of the replicate column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}

\item{species}{The name of the species column from data1}
}
\value{
output The stability of community abundance, calculated as mean/standard deviation
}
\description{
A function to calculate community stability over time
}

