% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temporal_torus_translation.R
\name{confint.cyclic_shift}
\alias{confint.cyclic_shift}
\title{Confidence Intervals Using a Cyclic Shift Significance Testing}
\usage{
\method{confint}{cyclic_shift}(object, parm = "out", level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{cyclic_shift}}

\item{parm}{which parameter is to be given a confidence interval. At present there is only one option: the mean of the null distribution. Defaults to "out", referring to the null distribution in objects of class \code{cyclic_shift}.}

\item{level}{the confidence level required.}

\item{...}{further arguments to \code{quantile}}
}
\value{
A dataframe with the following columns:
\itemize{
 \item{lowerCI: }{A numeric column with the lowest confidence interval value.}
 \item{upperCI: }{A numeric column with the highest confidence interval value.}
 \item{nullMean: }{A numeric column with the average value of the specified test statistic when calculated on a null community.}
 \item{replicate.var: }{A column that has same name and type as the replicate.var column, if replication is specified.}
}
}
\description{
Calculates confidence intervals for a user-specified test statistic that derives from a species x time matrix.
It does so by employing a cyclic shift that creates a null community by randomly selecting different starting points for each species' time series. This generates a community in which species abundances vary independently but within-species autocorrelation is maintained (Harms et al. 2001, Hallett et al. 2014).
This randomization is repeated a user-specific number of times and confidence intervals are reported for the resultant null distribution of the test statistic.
If the data frame includes multiple replicates, the test statistics for the null communities are averaged within each iteration unless specified otherwise.
}
\details{
This function was developed to be applied to the variance ratio; to use it for that purpose see the variance_ratio function.
It is included as a general function here but has very specific requirements. Namely, it is only relevant for functions that return single values, and for which varying species abundances independently is an acceptable way to develop a null test statistic value.

If applying this function to test statistics other the variance ratio, the input data frame needs to contain columns for time, species and abundance; time.var, species.var and abundance.var are used to indicate which columns contain those variables.
If multiple replicates are included in the data frame, that column should be specified with replicate.var. Each replicate should reflect a single experimental unit - there must be a single abundance value per species within each time point and replicate.
Null model confidence intervals default to the standard lowest 2.5\% and upper 97.5\% of the null distribution, typically these do not need to be change, but they can be user-modified to set more stringent CIs.
}
\references{
Hallett, Lauren M., Joanna S. Hsu, Elsa E. Cleland, Scott L. Collins, Timothy L. Dickson, Emily C. Farrer, Laureano A. Gherardi, et al. "Biotic Mechanisms of Community Stability Shift along a Precipitation Gradient." Ecology 95, no. 6 (2014): 1693-1700.

Harms, Kyle E., Richard Condit, Stephen P. Hubbell, and Robin B. Foster. "Habitat Associations of Trees and Shrubs in a 50-Ha Neotropical Forest Plot." Journal of Ecology 89, no. 6 (2001): 947-59.
}

