% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{get_lags}
\alias{get_lags}
\title{Get the lags from a distance matrix.}
\usage{
get_lags(DM, comm_data)
}
\arguments{
\item{DM}{distance matrix to be used for lag calculations}

\item{comm_data}{community data frame}
}
\value{
matrix of lag values
}
\description{
Get the lags from a distance matrix.
}

