% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/timelag_analysis.R
\name{get_lags}
\alias{get_lags}
\title{Get the lags from a distance matrix.}
\usage{
get_lags(DM)
}
\arguments{
\item{DM}{distance matrix to be used for lag calculations}
}
\value{
matrix of lag values
}
\description{
Get the lags from a distance matrix.
}

