% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/codyn-package.r
\docType{data}
\name{knz_001d}
\alias{knz_001d}
\title{Data from Konza Prairie, watershed 001d}
\format{A data frame with 8768 rows and 4 variables}
\usage{
data(knz_001d)
}
\description{
Plant composition within multiple replicates at an annually burned tallgrass
prairie site in the Konza Prairie LTER, Manhattan KS (Watershed 001D).
}
\details{
species,year,subplot,abundance
\itemize{
  \item species The sampled species
  \item year The sampling frequency
  \item subplot The sampled subplot
  \item abundance The measure of species abundance
}
}
\keyword{datasets}

