% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/rankshifts.R
\name{meanrank}
\alias{meanrank}
\title{Function for calculating mean rank shifts}
\usage{
meanrank(comm_data = dat, year = "year", species = "species",
  abundance = "abundance")
}
\arguments{
\item{comm_data}{Community dataset. Must be in 'long' format.}

\item{timevar}{The time variable}

\item{speciesvar}{The species variable}

\item{abdvar}{The abundance variable}
}
\value{
a dataframe, showing years compared
}
\description{
This is a function that calculates mean rank shifts
}

