% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{meanrank}
\alias{meanrank}
\title{Function for calculating mean rank shifts}
\usage{
meanrank(comm_data = dat, timevar = "year", speciesvar = "species",
  abdvar = "abundance")
}
\arguments{
\item{comm_data}{Community dataset. Must be in 'long' format.}

\item{timevar}{The time variable}

\item{speciesvar}{The species variable}

\item{abdvar}{The abundance variable}
}
\value{
a dataframe, showing years compared
}
\description{
This is a function that calculates mean rank shifts
}

