% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankshifts.R
\name{meanrankshift}
\alias{meanrankshift}
\title{Function for calculating mean rank shifts}
\usage{
meanrankshift(data1, replicate = "replicate", year = "year",
  species = "species", abundance = "abundance")
}
\arguments{
\item{data1}{Community dataset. Must be in 'long' format.}

\item{replicate}{The replication variable}

\item{year}{The time variable}

\item{species}{The species variable}

\item{abundance}{The abundance variable}
}
\value{
a dataframe, showing years compared
}
\description{
This is a function that calculates mean rank shifts
}

