% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankshifts.R
\name{meanrankshift}
\alias{meanrankshift}
\title{Function for calculating mean rank shifts}
\usage{
meanrankshift(df, time.var = "year", species.var = "species",
  abundance.var = "abundance", replicate.var = as.character(NA))
}
\arguments{
\item{df}{dataframe of Community dataset. Must be in 'long' format.}

\item{time.var}{The time variable}

\item{species.var}{The species variable}

\item{abundance.var}{The abundance variable}

\item{replicate.var}{The replication variable}
}
\value{
a dataframe, showing years compared
}
\description{
This is a function that calculates mean rank shifts
}

