% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_change.R
\name{multivariate_change}
\alias{multivariate_change}
\title{Multivariate changes in composition and dispersion}
\usage{
multivariate_change(df, time.var, species.var, abundance.var, replicate.var,
  treatment.var = NULL)
}
\arguments{
\item{df}{A data frame containing time, species, abundance and replicate columns and an optional column of treatment}

\item{time.var}{The name of the time column}

\item{species.var}{The name of the species column}

\item{abundance.var}{The name of the abundance column}

\item{replicate.var}{The name of the replicate column}

\item{treatment.var}{the neame of the optional treatment column}
}
\value{
The multivariate_change function returns a data frame with the following attributes:
\itemize{
 \item{time.var_pair: }{A characteric column that has the time points to be compared, separated by a dash.}
 \item{composition_change: }{A numeric column that is the euclidean distance between the centroids of two consecutive time points.}
 \item{dispersion_change: }{A numeric column that is the difference in the average dispersion of the replicates around the centriod for the two consecutive time periods. A negative value indicates replicates are converging over time (there is less dispersion at time peroid 2 than time period 1) and a postive value indicates replicates are diverging over time (there is more dispersion at time period 2 than time period 1.}
 \item{treatment.var: }{A column that has same name and type as the treatment.var column, if treatment.var is specified.}
}
}
\description{
Calculates the changes in composition and dispersion based off a Bray-Curtis dissimilarity matrix. Composition change is the euclidean distance between the centroids of time period 1 and time period 2. Dispersion change is the difference of average dispersion of each replicate to its centroid between consecutive time periods.
}
\examples{
data(pplots)
#With treatment
multivariate_change(pplots,
                    time.var="year", 
                    replicate.var = "plot", 
                    treatment.var = "treatment", 
                    species.var = "species", 
                    abundance.var = "relative_cover")

#Without treatment
df <- subset(pplots, treatment == "N1P0")
multivariate_change(df, 
                    time.var="year", 
                    replicate.var = "plot", 
                    species.var = "species", 
                    abundance.var = "relative_cover")
}
\references{
Avolio et al. 2015; Avolio et al. OUR PAPER, Mari Anderson?
}
