% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nullVRCI}
\alias{nullVRCI}
\title{A function to generate lower 2.5\% CI, upper 97.5\% CI and mean null VR values}
\usage{
nullVRCI(data1, species, year, abundance, bootnumber)
}
\arguments{
\item{data1}{A dataframe containing year, rep, species and abundance columns}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}

\item{bootnumber}{The number of null model iterations used to calculated CIs}
}
\value{
output A dataframe containing nullVRCIlow, nullVRCIhigh and nullVRmean
         nullVRCIow is the 0.025 CI and nullVRCIhigh is the 0.975 CI
}
\description{
A function to generate lower 2.5\% CI, upper 97.5\% CI and mean null VR values
}

