% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAC_change.R
\name{RAC_change}
\alias{RAC_change}
\title{Rank Abundance Curve Changes}
\usage{
RAC_change(df, time.var, species.var, abundance.var, replicate.var = NULL)
}
\arguments{
\item{df}{A data frame containing time, species, and abundance columns and an
optional column of replicates}

\item{time.var}{The name of the time column}

\item{species.var}{The name of the species column}

\item{abundance.var}{The name of the abundance column}

\item{replicate.var}{The name of the optional replicate column}
}
\value{
The RAC_change function returns a data frame with the following
  attributes:
\itemize{
 \item{replicate.var: }{A column that has same name and type as the
 replicate.var column, if replicate.var is specified.}
 \item{time.var_pair: }{A characteric column that has the time points to be
 compared, separated by a dash.}
 \item{richness_change: }{A numeric column that is the change in richness
 between the two consecutive time peroids for a repicate divided by the total
 number of species in both time periods.}
 \item{evenness_change: }{A numeric column that is the change in evenness
 (measured with EQ) between the two consecutive time peroids for a repicate
 divided by the total number of species in both time periods.}
 \item{rank_change: }{A numeric column that is the average change in rank of
 a species between the two consecutive time peroids for a repicate divided by
 the total number of species in both time periods. Species that are not
 present in both time periods are given the S+1 rank in the sample it is
 absent in, where S is the number of species in that sample.}
 \item{gains: }{A numeric column of the number of species that are present at
 time period 2 that were not present at time period 1 for a replicate divided
 by the total number of species in both time periods. This is equivelant to
 the turnover function with metric = "appearances".}
 \item{losses: }{A numeric column of the number of species that are not
 present at time period 2 but were  present at time period 1 for a replicate
 divided by the total number of species in both time periods. This is
 equivelant to the turnover function with metric = "disappearance".}
}
}
\description{
Calculates change of the five aspects of rank abundance curves
  (richness, evenness, rank, species gains, and species losses) for a
  replicate between two consecutive time points.
}
\examples{
data(pplots)
# Without replicates
df <- subset(pplots, plot == 25)
RAC_change(df = df,
           species.var = "species",
           abundance.var = "relative_cover",
           time.var = "year")

# With replicates
df <- subset(pplots, year < 2004 & plot \%in\% c(6, 25, 32))
RAC_change(df = df,
           species.var = "species",
           abundance.var = "relative_cover",
           replicate.var = "plot",
           time.var = "year")
}
\references{
Avolio et al.OUR PAPER
}
