% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{rate_change}
\alias{rate_change}
\title{Function quantifying the temporal rate and pattern of change in community composition.}
\usage{
rate_change(comm_data, year = "year")
}
\arguments{
\item{comm_data}{Community dataset. Must be in 'wide' format}

\item{year}{Year variable. Must be the first column of comm_data}
}
\value{
linear model coefficients
}
\description{
This is an analysis of differences in species composition between samples at increasing time lags. It measures the rate of directional change in community composition. First, a triangular dissimilarity matrix is calculated using Euclidean distance. Then, the Euclidean distance values are plotted against time lags. For example, a data set with 6 time intervals will have 5 one-year time lags (year 1 vs year 2, year 2 vs year 3 ...) and 4 two-year time lags (year 1 vs year 3, year 2 vs year 4 ...). Finally, distance values are regressed against time lag. The slope of the regression line indicates the rate and direction of change, and the regression coefficient is a measure of signal verses noise.
}

