% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{rate_change}
\alias{rate_change}
\title{A function to calculate community rate changes over time within multiple replicates}
\usage{
rate_change(data1, replicate = "replicate", year = "year",
  species = "species", abundance = "abundance")
}
\arguments{
\item{data1}{A dataframe containing replicate, year, species and abundance columns.}

\item{replicate}{The name of the replicate column from data1}

\item{year}{The name of the year column from data1}

\item{species}{The name of the species column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
output The rate of community change
}
\description{
#' This is an analysis of differences in species composition between samples at increasing time lags. It measures the rate of directional change in community composition. First, a triangular dissimilarity matrix is calculated using Euclidean distance. Then, the Euclidean distance values are plotted against time lags. For example, a data set with 6 time intervals will have 5 one-year time lags (year 1 vs year 2, year 2 vs year 3 ...) and 4 two-year time lags (year 1 vs year 3, year 2 vs year 4 ...). Finally, distance values are regressed against time lag. The slope of the regression line indicates the rate and direction of change.
}

