% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rate_change.R
\name{rate_change}
\alias{rate_change}
\title{A function to calculate community rate changes over time within multiple replicates}
\usage{
rate_change(df, time.var = "time", species.var = "species",
  abundance.var = "abundance", replicate.var = NA)
}
\arguments{
\item{df}{A dataframe containing replicate, time, species and abundance columns.}

\item{time.var}{The name of the time column from df}

\item{species.var}{The name of the species column from df}

\item{abundance.var}{The name of the abundance column from df}

\item{replicate.var}{The name of the replicate column from df. Defaults to NA.}
}
\value{
output The rate of community change
}
\description{
#' This is an analysis of differences in species composition between samples at increasing time lags. It measures the rate of directional change in community composition. First, a triangular dissimilarity matrix is calculated using Euclidean distance. Then, the Euclidean distance values are plotted against time lags. For example, a data set with 6 time intervals will have 5 one-year time lags (year 1 vs year 2, year 2 vs year 3 ...) and 4 two-year time lags (year 1 vs year 3, year 2 vs year 4 ...). Finally, distance values are regressed against time lag. The slope of the regression line indicates the rate and direction of change.
}

