% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_stability.R
\name{stability_onerep}
\alias{stability_onerep}
\title{A function to calculate species synchrony over time within one replicate}
\usage{
stability_onerep(df, x)
}
\arguments{
\item{df}{A dataframe containing x column}

\item{x}{The column to calculate stability on}
}
\value{
Stability of x, calculated as the mean/sd
}
\description{
A function to calculate species synchrony over time within one replicate
}
