% Generated by roxygen2 (4.0.2): do not edit by hand
\name{synchrony}
\alias{synchrony}
\title{A function to calculate species synchrony over time within multiple replicates}
\usage{
synchrony(data1, replicate = "replicate", species = "species",
  year = "year", abundance = "abundance")
}
\arguments{
\item{data1}{A dataframe containing replicate, year, species and abundance columns}

\item{replicate}{The name of the replicate column from data1}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}
}
\value{
output The degree of species synchrony, where 1 is perfect synchrony and 0 is perfect asynchrony
}
\description{
A function to calculate species synchrony over time within multiple replicates
}

