% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/synchrony.R
\name{synchrony}
\alias{synchrony}
\title{A function to calculate species synchrony over time within multiple replicates}
\usage{
synchrony(data1, replicate = "replicate", species = "species",
  year = "year", abundance = "abundance", metric = "Loreau")
}
\arguments{
\item{data1}{A dataframe containing replicate, year, species and abundance columns}

\item{replicate}{The name of the replicate column from data1}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}

\item{metric}{The synchrony metric to return. The default, "Loreau", returns synchrony as calculated by Loreau and de Mazancourt 2008.
The alternative, "Gross", returns synchrony as calculated by Gross et al. 2014}
}
\value{
output The degree of species synchrony. If "Loreau", 1 is perfect synchrony and 0 is perfect asynchrony.
       If "Gross", 1 is perfect synchrony and -1 is perfect asynchrony.
}
\description{
A function to calculate species synchrony over time within multiple replicates
}

