% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/synchrony.R
\name{synchrony}
\alias{synchrony}
\title{A function to calculate species synchrony over time within multiple replicates}
\usage{
synchrony(df, time.var = "year", species.var = "species",
  abundance.var = "abundance", metric = "Loreau", replicate.var = NA)
}
\arguments{
\item{df}{A dataframe containing replicate, time, species and abundance columns}

\item{time.var}{The name of the time column from df}

\item{species.var}{The name of the species column from df}

\item{abundance.var}{The name of the abundance column from df}

\item{metric}{The synchrony metric to return. The default, "Loreau", returns synchrony as calculated by Loreau and de Mazancourt 2008.
The alternative, "Gross", returns synchrony as calculated by Gross et al. 2014}

\item{replicate.var}{The name of the replicate column from df}
}
\value{
output The degree of species synchrony. If "Loreau", 1 is perfect synchrony and 0 is perfect asynchrony.
       If "Gross", 1 is perfect synchrony and -1 is perfect asynchrony.
}
\description{
A function to calculate species synchrony over time within multiple replicates
}

