% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temporal_torus_translation.R
\name{temporal_torus_translation}
\alias{temporal_torus_translation}
\title{Temporal Modification of the Torus Translation}
\usage{
temporal_torus_translation(df, time.var = "year", species.var = "species",
  abundance.var = "abundance", FUN)
}
\arguments{
\item{df}{A data frame containing time, species and abundance columns and an optional column of replicates}

\item{time.var}{The name of the time column}

\item{species.var}{The name of the species column}

\item{abundance.var}{The name of the abundance column}

\item{FUN}{A function to calculate on the null community}
}
\value{
The temporal_torus_translation function returns an integer output of the user-specified function calculated on a null community.
}
\description{
Calculates a user-specified test statistic on a null ecological community using a temporal modification of the Torus translation.
}
\details{
The input data frame needs to contain columns for time, species and abundance; time.var, species.var and abundance.var are used to indicate which columns contain those variables.
}

