% Generated by roxygen2 (4.0.2): do not edit by hand
\name{temporal_torus_translation}
\alias{temporal_torus_translation}
\title{A function to calculate a vector of null test statistics using a temporal modification of the Torus translation}
\usage{
temporal_torus_translation(data1, species, year, abundance, FUN)
}
\arguments{
\item{data1}{A dataframe containing year, species and abundance columns}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}

\item{FUN}{A function to calculate on the null community}
}
\value{
a vector of null test statistics calculated from a randomized community matrix in which species autocorrelation has been maintained via a Torus translation
}
\description{
A function to calculate a vector of null test statistics using a temporal modification of the Torus translation
}

