% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/temporal_torus_translation.R
\name{temporal_torus_translation_CI}
\alias{temporal_torus_translation_CI}
\title{A function that returns confidence intervals calculated from a temporal modification of the torus translation}
\usage{
temporal_torus_translation_CI(data1, replicate = "replicate",
  species = "species", year = "year", abundance = "abundance", FUN,
  bootnumber, li = 0.025, ui = 0.975, averagereps = TRUE)
}
\arguments{
\item{data1}{A dataframe containing year, species and abundance columns}

\item{replicate}{The name of the replication column from data1}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}

\item{FUN}{A function to calculate on the null community}

\item{bootnumber}{The number of null model iterations used to calculated CIs}

\item{li}{The lower confidence interval, defaults to lowest 2.5\% CI}

\item{ui}{The upper confidence interval, defaults to upper 97.5\% CI}

\item{averagereps}{If true returns CI averaged across reps; if false returns CI for each rep}
}
\value{
output A dataframe containing lowerCI, upperCI and nullmean value of the test statistic
}
\description{
A function that returns confidence intervals calculated from a temporal modification of the torus translation
}

