% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{transpose_community}
\alias{transpose_community}
\title{Convert from a longform abundance dataframe to a time by species dataframe.}
\usage{
transpose_community(df, time.var, species.var, abundance.var)
}
\arguments{
\item{df}{A dataframe containing time.var, species.var and abundance.var columns}

\item{time.var}{The name of the time column from df}

\item{species.var}{The name of the species column from df}

\item{abundance.var}{The name of the abundance column from df}
}
\value{
A dataframe of species abundances x time
}
\description{
Convert from a longform abundance dataframe to a time by species dataframe.
}
