% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/turnover.R
\name{turnover}
\alias{turnover}
\title{Species turnover}
\usage{
turnover(df, time.var = "year", species.var = "species",
  abundance.var = "abundance", replicate.var = as.character(NA),
  metric = "total")
}
\arguments{
\item{df}{A dataframe containing time.var, species.var and abundance.var columns and an optional replicate.var column}

\item{time.var}{The name of the time column}

\item{species.var}{The name of the species column}

\item{abundance.var}{The name of the abundance column}

\item{replicate.var}{The name of the replicate column}

\item{metric}{The turnover metric to return:
\itemize{
 \item{"total": }{The default metric, calculates summed appearances and disappearances relative to total species richness across both time periods}
 \item{"appearance": }{Calculates the number of species that appeared in the second time period relative to total species richness across both time periods}
 \item{"disappearance: }{Calculates the number of species that disappeared in the second time period relative to total species richness across both time periods}
}}
}
\value{
output A dataframe containing the specificed turnover metric and year
}
\description{
Computes species turnover between time periods as the proportion of species either gained or lost relative to the total number of species observed across both time periods.
Includes an option to compute turnover as just the proportion of species gained (i.e., "appearances") or lost (i.e., "disappearances").
}

