% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{turnover}
\alias{turnover}
\title{A function to calculate species turnover between years}
\usage{
turnover(data1, rep, species, year, abundance, metric = "turnover")
}
\arguments{
\item{data1}{A dataframe containing year,species and abundance columns}

\item{rep}{The name of the replicate column from data1}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}

\item{metric}{The turnover metric to return; the default, turnover, returns summed appearances and disappearances relative to total species richness across both years
appearance returns the number of appearances in the second year relative to total species richness across both years
disappearance returns the number of disappearances in the second year relative to the total species richness across both years}
}
\value{
output A dataframe containing the specificed turnover metric and year
}
\description{
A function to calculate species turnover between years
}

