% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{VR}
\alias{VR}
\title{A function to calculate the variance ratio and null model mean and CIs within multiple replicates}
\usage{
VR(data1, rep = "rep", species = "species", year = "year",
  abundance = "abundance", bootnumber, averagereps = TRUE)
}
\arguments{
\item{data1}{A dataframe containing rep, species, year and abundance columns}

\item{rep}{The name of the replicate column from data1}

\item{species}{The name of the species column from data1}

\item{year}{The name of the year column from data1}

\item{abundance}{The name of the abundance column from data1}

\item{bootnumber}{The number of null model iterations used to calculated CIs}

\item{averagereps}{If true returns VR and CI averaged across reps; if false returns VR and CI for each rep
If true, null VR are calculated within each rep, averaged, and the repeated for length of bootnumber}
}
\value{
output A dataframe containing the replicate name, VR  nullVRCIlow, nullVRCIhigh and nullVRmean
         VR is the actual variance ratio
         nullVRCIow is the 0.025 CI
         nullVRCIhigh is the 0.975 CI
         nullVRmean is the mean variance ratio calculated on null communities
}
\description{
A function to calculate the variance ratio and null model mean and CIs within multiple replicates
}

